-redirect user after logout -add static hamburger menu in right corner - add hook with firebase auth functions and delete simple auth functions
