--- src/firebase/auth/index.ts
@@ -1,2 +1,0 @@
-export * from "./logOut";
-export * from "./loginWithGoogle";
